pyproject.toml
[tool.poetry]
name = "gradebookgc_backend"
version = "0.1.0"
description = ""
authors = ["Vitalik <vzaharkiv28@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
django-filter = "^21.1"
drf-yasg = "^1.20.0"
django-debug-toolbar = "^3.2.4"
django-cors-headers = "^3.11.0"
django-lifecycle = "^0.9.6"
pytest = "^7.0.1"
pytest-django = "^4.5.2"
django-ninja = "^0.17.0"
python-jose = "^3.3.0"
mixer = "^7.2.1"
gTTS = "^2.2.4"
pandas = "^1.4.2"
pytest-cov = "^3.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
